Enable summary XLSX file generation and fix var references
This was disabled with a comment stating that it was calculated in a different
way, but the dashboard only runs this R script.

We comment out `debtsheet` because loading it is also commented out in more
recent versions of this R script.

We only need this workbook for raw data so we removed code that added visual
elements any was raising errors.