Disable conditional formatting when generating final Excel output
The World Bank deprecated this Excel file and is creating a different workbook,
but our data processing still requires it. We were encountering exceptions when
running the script so, since we only need this as a data source and do not need
nice formatting, we simply commented out all of the conditional formatting
statements.